Rename guava packages to Yelp to ease gradle dependency woes
