+++ src/com/yelp/common/base/Objects.java
@@ -9,7 +9,7 @@
  * governing permissions and limitations under the License.
  */
 
-package com.google.common.base;
+package com.yelp.common.base;
 
 import java.util.Arrays;
