+++ src/com/yelp/common/collect/AbstractMapEntry.java
@@ -9,11 +9,11 @@
  * governing permissions and limitations under the License.
  */
 
-package com.google.common.collect;
+package com.yelp.common.collect;
 
 import java.util.Map.Entry;
 
-import com.google.common.base.Objects;
+import com.yelp.common.base.Objects;
 
 /**
  * Implementation of the {@code equals}, {@code hashCode}, and {@code toString}
